requirements.txt
# Temel Django paketleri
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-environ==0.11.2

# Veritabanı ve Önbellek
redis>=4.5.0
django-redis==5.3.0
psycopg2-binary==2.9.9
django-db-connection-pool==1.2.3

# Geliştirme Araçları
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
django-debug-toolbar==4.3.0
django-extensions==4.1.0

# Dağıtım ve Statik Dosyalar
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.29.3

# Görev Yönetimi
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Kimlik Doğrulama ve API
django-allauth==0.61.1
django-rest-swagger==2.2.0
dj-rest-auth==4.0.1

# Yapay Zeka ve Makine Öğrenmesi
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
prophet==1.1.4
pytesseract==0.3.10
opencv-python==4.8.1.78
plotly==5.18.0
joblib==1.3.2

# Güvenlik ve Yetkilendirme
python-jose==3.3.0
cryptography==42.0.2
django-ratelimit==4.1.0
django-audit-log==0.7.0

# Test ve Kod Kalitesi
coverage==7.3.2
factory-boy==3.3.0

# Diğer Paketler
python-dotenv==1.0.0
openai==1.6.1
web3==6.15.1
requests==2.31.0
psutil==5.9.8

# Masaüstü Uygulaması
SQLAlchemy==2.0.27
alembic==1.1.3

# PWA Support
django-pwa==1.1.0

# Channels
channels==4.0.0
daphne==4.0.0

# Development
django-widget-tweaks==1.4.12
isort==5.12.0

# Miscellaneous
Pillow==10.2.0
python-dateutil==2.8.2

# API Documentation
drf-yasg==1.21.7

# Entegrasyon Paketleri
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
shopify-python-api==12.3.0
woocommerce==3.0.0
magento-api-client==1.0.0
iyzipay==1.0.37
paytr==1.0.0
payu-python==1.0.0
zeep==4.2.1
xmltodict==0.13.0
redis==5.0.1
celery==5.3.4

# Hava Durumu ve Finans API'leri
requests-ip-rotator==1.0.14
yfinance==0.2.36
alpha_vantage==2.3.1
finnhub-python==2.4.18
openweathermap==0.1.0

# GİB Entegrasyonu
lxml==4.9.3
requests==2.31.0
python-dotenv==1.0.0
xmlsig==0.1.1

# Added from the code block
pandas==2.1.0
openpyxl==3.1.2
xlsxwriter==3.1.2
weasyprint==60.1
tabulator-tables==5.5.0
flatpickr==4.6.13
htmx.org==1.9.6

# Google Cloud Vision
google-cloud-vision==3.4.4
google-cloud-core==2.3.3
google-api-core==2.11.1
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-storage==2.10.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
proto-plus==1.22.3
protobuf==4.23.4

# Added from the code block
dj-database-url==2.1.0

# Monitoring ve Sağlık Kontrolleri
sentry-sdk==1.40.6
django-health-check==3.17.0

# Security Vulnerability Scanning
bandit